// src/types/montazh.ts
export type ClipType = "a-roll" | "b-roll" | "synth";
export type TransitionType =
	| "cut"
	| "crossfade"
	| "dissolve"
	| "fade-to-black"
	| "fade-to-white"
	| "wipe-left"
	| "wipe-right"
	| "slide-left"
	| "slide-right";
// ...
export interface PaperEditClip {
	clip_id: string;
	source_file: string;
	in_ms: number;
	out_ms: number;
	clip_type: ClipType;
	overlay_text?: string;
	transition?: TransitionType;
	transition_duration_ms?: number;
	search_query?: string;
	description: string;
}